Final: Test containers Integration test
**Overview**
In this commit I started by adding the import dependencies to support the use of test containers and PostgreSQL for test containers . The use of testcontainers is the best practice because it makes sure that you dont need to use the real database or add multiple external dependencies in order to run our integration tests. This is done while respecting the need for the environment to be as close as possible to the real environment.

###What I did
- Added dependencies in the pom.xml file
-  Created a singleton container pattern to manage the containers and enable easy reuse and avoid mutation and redundancy.
- Wrote Integration tests using throwaway testcontainers.
**To-do**
A more advanced course on Spring boot testing. Maybe Junit in Action Book...